requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.0
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
cssselect==1.2.0
cssutils==2.11.1
frozenlist==1.5.0
geojson==2.5.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.5.0
lxml==5.3.0
more-itertools==10.5.0
multidict==6.1.0
phonenumbers==8.13.49
premailer==3.10.0
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
pyowm==3.3.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
SecretStorage==3.3.3
setuptools==75.3.0
soupsieve==2.6
twilio==9.3.6
urllib3==2.2.3
yagmail==0.15.293
yarl==1.17.1
